abstract observable from settings
